+++ nsplusthon/client/messages.py
@@ -1,3 +1,4 @@
+import asyncio
 import inspect
 import itertools
 import logging